requirements.txt
fastapi
uvicorn
gunicorn
python-dotenv
pymilvus
sentence-transformers
google-generativeai
beautifulsoup4
playwright
